tests: split descriptor tests from refinement tests
